Generalize update and fix hit test for rotated elements
