+++ src/element/dragElements.ts
@@ -1,17 +1,10 @@
-import {
-  NonDeletedExcalidrawElement,
-  ExcalidrawLinearElement,
-  NonDeleted,
-  PointBinding,
-  ExcalidrawBindableElement,
-} from "./types";
+import { SHAPES } from "../shapes";
+import { updateBoundElements } from "./binding";
 import { getCommonBounds } from "./bounds";
 import { mutateElement } from "./mutateElement";
-import { SHAPES } from "../shapes";
 import { getPerfectElementSize } from "./sizeHelpers";
-import { LinearElementEditor } from "./linearElementEditor";
 import Scene from "../scene/Scene";
-import { intersectElementWithLine, pointInAbsoluteCoords } from "./collision";
+import { NonDeletedExcalidrawElement } from "./types";
 
 export const dragSelectedElements = (
   selectedElements: NonDeletedExcalidrawElement[],
@@ -26,32 +19,7 @@ export const dragSelectedElements = (
       x: element.x + offset.x,
       y: element.y + offset.y,
     });
-    updateBoundElementsOnDrag(element, offset, scene);
-  });
-};
-
-const updateBoundElementsOnDrag = (
-  draggedElement: NonDeletedExcalidrawElement,
-  offset: { x: number; y: number },
-  scene: Scene,
-) => {
-  (scene.getNonDeletedElements(
-    draggedElement.boundElementIds ?? [],
-  ) as NonDeleted<ExcalidrawLinearElement>[]).forEach((boundElement) => {
-    maybeMoveBoundPoint(
-      boundElement,
-      "start",
-      boundElement.startBinding,
-      draggedElement as ExcalidrawBindableElement,
-      offset,
-    );
-    maybeMoveBoundPoint(
-      boundElement,
-      "end",
-      boundElement.endBinding,
-      draggedElement as ExcalidrawBindableElement,
-      offset,
-    );
+    updateBoundElements(element);
   });
 };
 
@@ -106,66 +74,4 @@ export const dragNewElement = (
       height: height,
     });
   }
-};
-
-const maybeMoveBoundPoint = (
-  boundElement: NonDeleted<ExcalidrawLinearElement>,
-  startOrEnd: "start" | "end",
-  binding: PointBinding | null | undefined,
-  draggedElement: ExcalidrawBindableElement,
-  offset: { x: number; y: number },
-): void => {
-  if (binding?.elementId === draggedElement.id) {
-    moveBoundPoint(boundElement, startOrEnd, binding, draggedElement, offset);
-  }
-};
-
-const moveBoundPoint = (
-  linearElement: NonDeleted<ExcalidrawLinearElement>,
-  startOrEnd: "start" | "end",
-  binding: PointBinding,
-  draggedElement: ExcalidrawBindableElement,
-  offset: { x: number; y: number },
-): void => {
-  const direction = startOrEnd === "start" ? -1 : 1;
-  const edgePointIndex = direction === -1 ? 0 : linearElement.points.length - 1;
-  const adjacentPointIndex = edgePointIndex - direction;
-  // The linear element was not originally pointing inside the bound shape,
-  // we use simple binding without focus points
-  if (binding.gap === 0) {
-    LinearElementEditor.movePointByOffset(
-      linearElement,
-      edgePointIndex,
-      offset,
-    );
-    return;
-  }
-  const adjacentPoint = LinearElementEditor.getPointAtIndexGlobalCoordinates(
-    linearElement,
-    adjacentPointIndex,
-  );
-  const draggedFocusPointAbsolute = pointInAbsoluteCoords(
-    draggedElement,
-    binding.focusPoint,
-  );
-  const intersections = intersectElementWithLine(
-    draggedElement,
-    adjacentPoint,
-    draggedFocusPointAbsolute,
-    binding.gap,
-  );
-  if (intersections.length === 0) {
-    // TODO: This should never happen, but it does due to scaling/rotating
-    return;
-  }
-  // Guaranteed to intersect because focusPoint is always inside the shape
-  const [nearIntersection] = intersections;
-  LinearElementEditor.movePoint(
-    linearElement,
-    edgePointIndex,
-    LinearElementEditor.pointFromAbsoluteCoords(
-      linearElement,
-      nearIntersection,
-    ),
-  );
 };
